еренёс всю логику в Tree и Сщддусешщт https://online.sbis.ru/opendoc.html?guid=99613fa8-d9dd-4a0e-b715-b634331b344b
