+++ Controls/_treeGrid/display/TreeGridCollection.ts
@@ -175,46 +175,9 @@ export default class TreeGridCollection<
 
     setExpandedItems(expandedKeys: CrudEntityKey[]): void {
         super.setExpandedItems(expandedKeys);
-        this.resetLastItem();
-        this.resetFirstItem();
     }
 
     // endregion
-
-    // region Аспект "крайние записи"
-
-    getLastItem(): EntityModel {
-        if (!!this._lastItem) {
-            return this._lastItem;
-        }
-        return this._lastItem = this._getLastItemRecursive(this.getRoot().getContents());
-    }
-
-    private _getLastItemRecursive(root: S): S {
-        if (root && (!root.getKeyProperty() || root.getKey() === undefined)) {
-            return root;
-        }
-        // Обращаемся к иерархии для получения детей
-        const children = this._getChildrenByRecordSet(root);
-        const lastChild: S = children[children.length - 1];
-        // Если узел и у него нет детей, то он последний
-        if (children.length === 0) {
-            return root;
-        }
-        const isNode = lastChild.get(this._$nodeProperty) !== null;
-
-        // expandedItems появляются только после того, как был вызван Tree.setExpandedItems
-        const expandedItems = this.getExpandedItems();
-        if (isNode && expandedItems && (
-            this.isExpandAll() ||
-            (expandedItems && expandedItems.indexOf(lastChild.getKey() !== -1))
-        )) {
-            return this._getLastItemRecursive(lastChild);
-        }
-        return lastChild;
-    }
-
-    // endregion Аспект "крайние записи"
 
     // region HasNodeWithChildren
 
@@ -234,8 +197,6 @@ export default class TreeGridCollection<
         options.colspanCallback = this._$colspanCallback;
         options.columnSeparatorSize = this._$columnSeparatorSize;
         options.rowSeparatorSize = this._$rowSeparatorSize;
-        options.isLastItem = this.isLastItem(options.contents);
-        options.isFirstItem = this.isFirstItem(options.contents);
 
         // Строит обычную фабрику
         const CollectionItemsFactory = (factoryOptions?: ITreeGridRowOptions<S>): ItemsFactory<T> => {
